Refactor EndpointManager to HandlerManager
There's no need for multiple subcless of HandlerManager if the Handlers
can have their own subtype that knows how they get called. The manager
can detect the proper handler based on some common interface.

As a result, it's possible to mix different types of handlers in an
application if needed instead of being fixed to a single type.